refactor: [move] entity, repository로 각각 이동
