+++ src/main/java/site/travellaboratory/be/comment/application/CommentReaderService.java
@@ -14,8 +14,8 @@ import site.travellaboratory.be.common.exception.ErrorCodes;
 import site.travellaboratory.be.comment.infrastructure.persistence.repository.CommentJpaRepository;
 import site.travellaboratory.be.comment.infrastructure.persistence.entity.CommentEntity;
 import site.travellaboratory.be.comment.domain.enums.CommentStatus;
-import site.travellaboratory.be.review.infrastructure.persistence.ReviewJpaRepository;
-import site.travellaboratory.be.review.infrastructure.persistence.ReviewEntity;
+import site.travellaboratory.be.review.infrastructure.persistence.repository.ReviewJpaRepository;
+import site.travellaboratory.be.review.infrastructure.persistence.entity.ReviewEntity;
 import site.travellaboratory.be.review.domain.enums.ReviewStatus;
 import site.travellaboratory.be.user.infrastructure.persistence.repository.UserJpaRepository;
 import site.travellaboratory.be.user.infrastructure.persistence.entity.UserEntity;
