+++ src/main/java/site/travellaboratory/be/review/application/ReviewLikeService.java
@@ -11,9 +11,9 @@ import site.travellaboratory.be.review.domain.Review;
 import site.travellaboratory.be.review.domain.ReviewLike;
 import site.travellaboratory.be.review.domain.enums.ReviewLikeStatus;
 import site.travellaboratory.be.review.domain.enums.ReviewStatus;
-import site.travellaboratory.be.review.infrastructure.persistence.ReviewJpaRepository;
-import site.travellaboratory.be.review.infrastructure.persistence.ReviewLikeJpaRepository;
-import site.travellaboratory.be.review.infrastructure.persistence.ReviewLikeEntity;
+import site.travellaboratory.be.review.infrastructure.persistence.repository.ReviewJpaRepository;
+import site.travellaboratory.be.review.infrastructure.persistence.repository.ReviewLikeJpaRepository;
+import site.travellaboratory.be.review.infrastructure.persistence.entity.ReviewLikeEntity;
 import site.travellaboratory.be.user.infrastructure.persistence.repository.UserJpaRepository;
 import site.travellaboratory.be.user.infrastructure.persistence.entity.UserEntity;
 import site.travellaboratory.be.user.domain.enums.UserStatus;
