+++ src/main/java/site/travellaboratory/be/review/application/ReviewReaderService.java
@@ -10,14 +10,14 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import site.travellaboratory.be.common.exception.BeApplicationException;
 import site.travellaboratory.be.common.exception.ErrorCodes;
-import site.travellaboratory.be.review.infrastructure.persistence.ReviewJpaRepository;
-import site.travellaboratory.be.review.infrastructure.persistence.ReviewEntity;
+import site.travellaboratory.be.review.infrastructure.persistence.repository.ReviewJpaRepository;
+import site.travellaboratory.be.review.infrastructure.persistence.entity.ReviewEntity;
 import site.travellaboratory.be.review.domain.enums.ReviewStatus;
 import site.travellaboratory.be.user.infrastructure.persistence.repository.UserJpaRepository;
 import site.travellaboratory.be.user.infrastructure.persistence.entity.UserEntity;
 import site.travellaboratory.be.user.domain.enums.UserStatus;
-import site.travellaboratory.be.review.infrastructure.persistence.ReviewLikeJpaRepository;
-import site.travellaboratory.be.review.infrastructure.persistence.ReviewLikeEntity;
+import site.travellaboratory.be.review.infrastructure.persistence.repository.ReviewLikeJpaRepository;
+import site.travellaboratory.be.review.infrastructure.persistence.entity.ReviewLikeEntity;
 import site.travellaboratory.be.review.domain.enums.ReviewLikeStatus;
 import site.travellaboratory.be.review.presentation.response.reader.ProfileReviewLocation;
 import site.travellaboratory.be.review.presentation.response.reader.ProfileReviewPaginationResponse;
