+++ src/main/java/site/travellaboratory/be/review/application/ReviewWriterService.java
@@ -15,8 +15,8 @@ import site.travellaboratory.be.user.domain.enums.UserStatus;
 import site.travellaboratory.be.user.domain.User;
 import site.travellaboratory.be.article.infrastructure.persistence.repository.ArticleJpaRepository;
 import site.travellaboratory.be.article.infrastructure.persistence.entity.ArticleEntity;
-import site.travellaboratory.be.review.infrastructure.persistence.ReviewEntity;
-import site.travellaboratory.be.review.infrastructure.persistence.ReviewJpaRepository;
+import site.travellaboratory.be.review.infrastructure.persistence.entity.ReviewEntity;
+import site.travellaboratory.be.review.infrastructure.persistence.repository.ReviewJpaRepository;
 import site.travellaboratory.be.user.infrastructure.persistence.entity.UserEntity;
 import site.travellaboratory.be.user.infrastructure.persistence.repository.UserJpaRepository;
 import site.travellaboratory.be.review.domain.request.ReviewSaveRequest;
