+++ src/main/java/site/travellaboratory/be/review/infrastructure/persistence/repository/ReviewLikeJpaRepository.java
@@ -1,10 +1,11 @@
-package site.travellaboratory.be.review.infrastructure.persistence;
+package site.travellaboratory.be.review.infrastructure.persistence.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import site.travellaboratory.be.review.domain.enums.ReviewLikeStatus;
+import site.travellaboratory.be.review.infrastructure.persistence.entity.ReviewLikeEntity;
 
 public interface ReviewLikeJpaRepository extends JpaRepository<ReviewLikeEntity, Long> {
