+++ src/main/java/site/travellaboratory/be/review/presentation/response/reader/ProfileReviewResponse.java
@@ -2,7 +2,7 @@ package site.travellaboratory.be.review.presentation.response.reader;
 
 import java.time.LocalDate;
 import java.util.List;
-import site.travellaboratory.be.review.infrastructure.persistence.ReviewEntity;
+import site.travellaboratory.be.review.infrastructure.persistence.entity.ReviewEntity;
 
 public record ProfileReviewResponse(
     Long reviewId,
