+++ src/main/java/site/travellaboratory/be/review/presentation/response/reader/ReviewBannerResponse.java
@@ -1,7 +1,7 @@
 package site.travellaboratory.be.review.presentation.response.reader;
 
 import java.util.List;
-import site.travellaboratory.be.review.infrastructure.persistence.ReviewEntity;
+import site.travellaboratory.be.review.infrastructure.persistence.entity.ReviewEntity;
 
 public record ReviewBannerResponse(
     Long reviewId,
